tag binaries in msgpack to ensure transparency
This ensures that if you send a msgpack string or binary through erlang
things it comes out the same on the other side. To achieve this we
either need to tag strings or binaries, and for the most part binaries
generated in erlang will tend to be human-readable strings, so tagging
msgpack binaries makes more sense.